backend/requirements.txt
flask==3.1.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
requests==2.33.0
python-dotenv==1.0.0
gunicorn==21.2.0
